Second

+++ Task1/Program.cs
@@ -10,7 +10,7 @@ namespace Task1
     {
         static void Main(string[] args)
         {
-            int num1 = 10;
+            short num1 = 10;
             int num2 = 15;
 
             Console.WriteLine($"{num1}+{num2}={Program.Sum(num1,num2)}");
